feat : reservation detail service
예약한 정보를 불러온다.

+++ haul-be/src/main/java/com/hansalchai/haul/reservation/controller/ReservationRestController.java
@@ -67,7 +67,7 @@ public class ReservationRestController {
 	}
 
 	@GetMapping("/reservations/{id}")
-	public ResponseEntity<ApiResponse<ReservationDetailDTO>> getCustomerReservationDetail(@PathVariable int id, HttpServletRequest request){
+	public ResponseEntity<ApiResponse<ReservationDetailDTO>> getCustomerReservationDetail(@PathVariable Long id, HttpServletRequest request){
 		AuthenticatedUser auth = (AuthenticatedUser)request.getAttribute(AUTHENTICATE_USER);
 		ReservationDetailDTO response = reservationService.getReservationDetail(id, auth.getUserId());
 		return ResponseEntity.ok(success(SuccessCode.GET_SUCCESS, response));
